--- a/src/common/legacy_config_opts.h
+++ b/src/common/legacy_config_opts.h
@@ -180,7 +180,6 @@ OPTION(ms_dpdk_hw_flow_control, OPT_BOOL)
 // Weighing of a hardware network queue relative to a software queue (0=no work, 1=     equal share)")
 OPTION(ms_dpdk_hw_queue_weight, OPT_FLOAT)
 OPTION(ms_dpdk_debug_allow_loopback, OPT_BOOL)
-OPTION(ms_dpdk_rx_buffer_count_per_core, OPT_INT)
 
 OPTION(inject_early_sigterm, OPT_BOOL)